Add files via upload
This particular dashboard is different because it employs the use of "Smart Narratives" almost throughout the report. Smart Narratives give quick text summaries about a selected item on a PowerBI report. The SNAs summaries are effective in bringing to light salient points, and highlighting trends, especially in percentages.  These SNAs are fully customizable and can be attached to a lot of visuals to give a further summary and analysis.